Refactor delete bitmap calculator
